TW-51276 Read only: introduce ServerResponsibility

+++ server/src/jetbrains/buildServer/swabra/OldHandleToolConverter.java
@@ -18,10 +18,8 @@ package jetbrains.buildServer.swabra;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.UncheckedIOException;
-import jetbrains.buildServer.serverSide.SecurityContextEx;
 import jetbrains.buildServer.serverSide.ServerPaths;
-import jetbrains.buildServer.serverSide.auth.Permission;
+import jetbrains.buildServer.serverSide.ServerResponsibility;
 import jetbrains.buildServer.tools.ServerToolPreProcessorAdapter;
 import jetbrains.buildServer.tools.ToolException;
 import jetbrains.buildServer.tools.installed.ToolPaths;
@@ -39,14 +37,14 @@ public class OldHandleToolConverter extends ServerToolPreProcessorAdapter {
 
   private final ServerPaths myServerPaths;
   @NotNull private final ToolPaths myToolPaths;
-  @NotNull private final SecurityContextEx mySecurityContext;
+  @NotNull private final ServerResponsibility myServerResponsibility;
 
   public OldHandleToolConverter(@NotNull final ServerPaths serverPaths,
                                 @NotNull final ToolPaths toolPaths,
-                                @NotNull final SecurityContextEx securityContext) {
+                                @NotNull final ServerResponsibility serverResponsibility) {
     myServerPaths = serverPaths;
     myToolPaths = toolPaths;
-    mySecurityContext = securityContext;
+    myServerResponsibility = serverResponsibility;
   }
 
   @NotNull
@@ -57,44 +55,40 @@ public class OldHandleToolConverter extends ServerToolPreProcessorAdapter {
 
   @Override
   public void doBeforeServerStartup() throws ToolException {
-    try {
-      mySecurityContext.runIfHasGlobalPermission(() -> {
-        final File oldPlugin1 = new File(myServerPaths.getPluginsDir(), "handle-provider");
-        if (oldPlugin1.exists()) {
-          LOG.debug("Detected old handle-provider plugin " + oldPlugin1);
-          try {
-            final File validExeToolLocation = myToolPaths.getSharedToolPath(new File("handle.exe"));
-            if (!validExeToolLocation.isFile()) {
-              LOG.debug("Converting old handle-provider plugin " + oldPlugin1 + " into tool");
-              final File agentPlugin = new File(oldPlugin1, "agent/handle-provider.zip");
-              if (!agentPlugin.isFile()) {
-                LOG.warn("No agent plugin detected in " + oldPlugin1);
-              } else {
-                final File temp = new File(FileUtil.getTempDirectory(), "handle-provider");
-                try {
-                  ArchiveUtil.unpackZip(agentPlugin, "", temp);
-                  final File oldHandleExe = new File(temp, "handle-provider/bin/handle.exe");
-                  if (oldHandleExe.isFile()) {
-                    FileUtil.copy(oldHandleExe, validExeToolLocation);
-                  } else {
-                    LOG.warn("No handle.exe detected in " + oldPlugin1);
-                  }
-                } catch (IOException e) {
-                  throw new UncheckedIOException("Failed to extract handle.exe from " + agentPlugin, e);
-                } finally {
-                  FileUtil.delete(temp);
+    if (myServerResponsibility.canManageServerConfig()) {
+      final File oldPlugin1 = new File(myServerPaths.getPluginsDir(), "handle-provider");
+      if (oldPlugin1.exists()) {
+        LOG.debug("Detected old handle-provider plugin " + oldPlugin1);
+        try {
+          final File validExeToolLocation = myToolPaths.getSharedToolPath(new File("handle.exe"));
+          if (!validExeToolLocation.isFile()) {
+            LOG.debug("Converting old handle-provider plugin " + oldPlugin1 + " into tool");
+            final File agentPlugin = new File(oldPlugin1, "agent/handle-provider.zip");
+            if (!agentPlugin.isFile()) {
+              LOG.warn("No agent plugin detected in " + oldPlugin1);
+            } else {
+              final File temp = new File(FileUtil.getTempDirectory(), "handle-provider");
+              try {
+                ArchiveUtil.unpackZip(agentPlugin, "", temp);
+                final File oldHandleExe = new File(temp, "handle-provider/bin/handle.exe");
+                if (oldHandleExe.isFile()) {
+                  FileUtil.copy(oldHandleExe, validExeToolLocation);
+                } else {
+                  LOG.warn("No handle.exe detected in " + oldPlugin1);
                 }
+              } catch (IOException e) {
+                throw new ToolException("Failed to extract handle.exe from " + agentPlugin, e);
+              } finally {
+                FileUtil.delete(temp);
               }
             }
-          } finally {
-            LOG.debug("Deleting old handle-provider plugin " + oldPlugin1);
-            FileUtil.delete(oldPlugin1);
           }
+        } finally {
+          LOG.debug("Deleting old handle-provider plugin " + oldPlugin1);
+          FileUtil.delete(oldPlugin1);
         }
-        FileUtil.delete(new File(myServerPaths.getPluginDataDirectory(), "handle-provider.zip"));
-      }, Permission.CHANGE_SERVER_SETTINGS);
-    } catch (UncheckedIOException e) {
-      new ToolException(e.getMessage(), e);
+      }
+      FileUtil.delete(new File(myServerPaths.getPluginDataDirectory(), "handle-provider.zip"));
     }
   }
 }
